requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aioredis==2.0.1
aioschedule==0.5.2
aiosignal==1.3.1
annotated-types==0.6.0
async==0.6.2
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.0
frozenlist==1.4.0
google-api-core==2.12.0
google-api-python-client==2.105.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
greenlet==3.0.0
httplib2==0.22.0
idna==3.4
magic-filter==1.0.12
multidict==6.0.4
oauthlib==3.2.2
protobuf==4.24.4
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg-pool==3.1.8
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.3.0
pydantic_core==2.6.3
pyparsing==3.1.1
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
SQLAlchemy==2.0.21
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
yarl==1.9.2
